refactor(nav): extract TransitionController from ExperimentalControls
Move the preset-pose camera motions (pop-to-roof/daylight, swoop-to-street,
rise-to-drone, fall-to), the context-view-button resolver + dispatch
(resolveContextAction / triggerContextAction), and the ActionBar zoom into
transitionController.js. Each preset keeps its start->end pose math and hands the
tween to ctx.runner.run; the resolver is a pure read of the situation sensor's
context snapshot.

The orchestrator keeps resolveContextAction / triggerContextAction as public
delegators (the React context button + Space route through them); zoomInStart /
zoomOutStart drive ctx.transition.zoomActionBar; the recovery hand-off reaches
ctx.transition.popToRoof (public for that reason). ctx gains zoomSpeed,
minSpeedFactor, transition.

Behaviour-preserving: 302/302 characterization tests green.
